+++ src/BDSOutput.cc
@@ -103,7 +103,9 @@ BDSOutput::BDSOutput(const G4String& baseFileNameIn,
   energyDepositedWorldContents(0),
   energyDepositedTunnel(0),
   energyImpactingAperture(0),
+  energyImpactingApertureKinetic(0),
   energyWorldExit(0),
+  energyWorldExitKinetic(0),
   nCollimatorsInteracted(0)
 {
   const BDSGlobalConstants* g = BDSGlobalConstants::Instance();
@@ -300,7 +302,9 @@ void BDSOutput::FillEvent(const BDSEventInfo*                            info,
   energyDepositedWorldContents = 0;
   energyDepositedTunnel        = 0;
   energyImpactingAperture      = 0;
+  energyImpactingApertureKinetic = 0;
   energyWorldExit              = 0;
+  energyWorldExitKinetic       = 0;
   nCollimatorsInteracted       = 0;
   
   if (vertex && storePrimaries)
@@ -622,7 +626,9 @@ void BDSOutput::FillEventInfo(const BDSEventInfo* info)
   evtInfo->energyDepositedWorldContents = energyDepositedWorldContents;
   evtInfo->energyDepositedTunnel        = energyDepositedTunnel;
   evtInfo->energyWorldExit              = energyWorldExit;
+  evtInfo->energyWorldExitKinetic       = energyWorldExitKinetic;
   evtInfo->energyImpactingAperture      = energyImpactingAperture;
+  evtInfo->energyImpactingApertureKinetic = energyImpactingApertureKinetic;
   G4double ek = BDSStackingAction::energyKilled / CLHEP::GeV;
   evtInfo->energyKilled = ek;
   evtInfo->energyTotal =  energyDeposited
@@ -700,6 +706,7 @@ void BDSOutput::FillEnergyLoss(const BDSHitsCollectionEnergyDepositionGlobal* hi
 	  {
 	    BDSHitEnergyDepositionGlobal* hit = (*hits)[i];
 	    energyWorldExit += hit->TotalEnergyWeighted()/CLHEP::GeV;
+	    energyWorldExitKinetic += hit->KineticEnergyWeighted()/CLHEP::GeV;
 	    if (storeELossWorld)
 	      {eLossWorldExit->Fill(hit);}
 	  }
@@ -931,6 +938,7 @@ void BDSOutput::FillApertureImpacts(const BDSHitsCollectionApertureImpacts* hits
       const BDSHitApertureImpact* hit = (*hits)[i];
       G4double eW = (hit->totalEnergy / CLHEP::GeV) * hit->weight;
       energyImpactingAperture += eW;
+      energyImpactingApertureKinetic += (hit->preStepKineticEnergy * hit->weight ) / CLHEP::GeV;
       if (hit->parentID == 0)
 	{
 	  nPrimaryImpacts += 1;